fix: make Telegram port compile — JNI_FUNC macro, missing includes, type mismatches, missing JNI declarations, manifest entries, stub fixes
- C++: rename JNI_TG→JNI_FUNC in tg_jni_bridge.cpp (137 call sites fixed)
- C++: add #include <cctype> to tdlib_client.cpp (isspace)
- C++: add #include <algorithm> to tdlib_cache.cpp (std::reverse)
- C++: fix TlField aggregate init in tdlib_auth.cpp (isString=true via str())
- C++: remove duplicate reaction_types in tdlib_messages.cpp
- C++: fix 6 wrong method calls in tg_jni_bridge.cpp (sendAnimation, sendAudio, etc.)
- C++: add 8 missing wrapper methods to TdNativeClient
- C++: remove 4 dead callback JNI entry points
- Kotlin: fix _userId visibility (private→internal)
- Kotlin: fix Long/Int type mismatch (photoId), String/Int (Video.height)
- Kotlin: add 13 missing external fun declarations in ProgressiveNative.kt
- Manifest: add 3 missing activity declarations

+++ vector/src/main/java/chat/progressive/app/telegram/TelegramProfileService.kt
@@ -528,7 +528,7 @@ class TelegramProfileService(private val client: TdLibClient) {
                 val s = arr.getJSONObject(i)
                 TelegramPhoto.PhotoSize(
                     type = s.optString("type", ""),
-                    photoId = s.optInt("id", 0),
+                    photoId = s.optLong("id", 0L),
                     width = s.optInt("width", 0),
                     height = s.optInt("height", 0),
                     fileId = s.optJSONObject("photo")?.optInt("id", 0) ?: 0,
